make UserDAO to interface
