+++ BE/Spring/toby spring/Vol.1/springbook/src/main/java/org/example/user/dao/DaoFactory.java
@@ -1,6 +1,5 @@
 package org.example.user.dao;
 
-import com.mysql.cj.jdbc.exceptions.ConnectionFeatureNotAvailableException;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.datasource.SimpleDriverDataSource;
@@ -10,8 +9,8 @@ import javax.sql.DataSource;
 @Configuration
 public class DaoFactory {
     @Bean
-    public UserDAO userDAO(){
-        UserDAO userDAO = new UserDAO();
+    public UserDAOJdbc userDAO(){
+        UserDAOJdbc userDAO = new UserDAOJdbc();
         userDAO.setDataSource(dataSource());
         return userDAO;
     }
