+++ BE/Spring/toby spring/Vol.1/springbook/src/main/java/org/example/user/dao/UserDAO.java
@@ -1,57 +1,13 @@
 package org.example.user.dao;
 
-import com.mysql.cj.result.Row;
-import lombok.NoArgsConstructor;
 import org.example.user.domain.User;
-import org.springframework.dao.DataAccessException;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
-import org.springframework.jdbc.core.ResultSetExtractor;
-import org.springframework.jdbc.core.RowMapper;
 
-import javax.sql.DataSource;
-import java.sql.*;
 import java.util.List;
 
-@NoArgsConstructor
-public class UserDAO {
-    private JdbcTemplate jdbcTemplate;
-
-    private RowMapper<User> userMapper = new RowMapper<User>() {
-        @Override
-        public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-            User user = new User();
-            user.setId(rs.getString("id"));
-            user.setPassword(rs.getString("password"));
-            user.setName(rs.getString("name"));
-
-            return user;
-        }
-    };
-
-    public void setDataSource(DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
-    }
-
-    public void add(final User user) throws SQLException {
-        this.jdbcTemplate.update("insert into users(id, name, password) values (?,?,?)",
-                user.getId(), user.getName(), user.getPassword());
-    }
-
-    public User get(String id) throws SQLException {
-        return this.jdbcTemplate.queryForObject("select * from users where id = ?", new Object[]{id}, userMapper);
-    }
-
-    public void deleteAll() throws SQLException {
-        this.jdbcTemplate.update("delete from users");
-    }
-
-    public int getCount() throws SQLException {
-        return this.jdbcTemplate.queryForObject("select count(*) from users", Integer.class);
-    }
-
-    public List<User> getAll(){
-        return this.jdbcTemplate.query("select * from users order by id", userMapper);
-    }
+public interface UserDAO {
+    void add(User user);
+    User get(String id);
+    List<User> getAll();
+    void deleteAll();
+    int getCount();
 }
