+++ BE/Spring/toby spring/Vol.1/springbook/src/main/java/org/example/user/dao/UserDAOJdbc.java
@@ -1,0 +1,52 @@
+package org.example.user.dao;
+
+import lombok.NoArgsConstructor;
+import org.example.user.domain.User;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+
+import javax.sql.DataSource;
+import java.sql.*;
+import java.util.List;
+
+@NoArgsConstructor
+public class UserDAOJdbc implements UserDAO {
+    private JdbcTemplate jdbcTemplate;
+
+    private RowMapper<User> userMapper = new RowMapper<User>() {
+        @Override
+        public User mapRow(ResultSet rs, int rowNum) throws SQLException {
+            User user = new User();
+            user.setId(rs.getString("id"));
+            user.setPassword(rs.getString("password"));
+            user.setName(rs.getString("name"));
+
+            return user;
+        }
+    };
+
+    public void setDataSource(DataSource dataSource) {
+        this.jdbcTemplate = new JdbcTemplate(dataSource);
+    }
+
+    public void add(final User user){
+        this.jdbcTemplate.update("insert into users(id, name, password) values (?,?,?)",
+                user.getId(), user.getName(), user.getPassword());
+    }
+
+    public User get(String id){
+        return this.jdbcTemplate.queryForObject("select * from users where id = ?", new Object[]{id}, userMapper);
+    }
+
+    public void deleteAll(){
+        this.jdbcTemplate.update("delete from users");
+    }
+
+    public int getCount(){
+        return this.jdbcTemplate.queryForObject("select count(*) from users", Integer.class);
+    }
+
+    public List<User> getAll(){
+        return this.jdbcTemplate.query("select * from users order by id", userMapper);
+    }
+}
