+++ BE/Spring/toby spring/Vol.1/springbook/src/test/resources/test-applicationContext.xml
@@ -10,7 +10,7 @@
         <property name="password" value="root"/>
     </bean>
 
-    <bean id="userDAO" class="org.example.user.dao.UserDAO">
+    <bean id="userDAO" class="org.example.user.dao.UserDAOJdbc">
         <property name="dataSource" ref="dataSource"/>
     </bean>
 </beans>
